komunikaty i obsługa błędów na które się natknąłem

+++ BatchKmlToKmz/Form1.cs
@@ -9,6 +9,7 @@ using System.Windows.Forms;
 using System.IO;
 using System.Diagnostics;
 
+
 namespace WindowsFormsApplication1
 {
     public partial class Form1 : Form
@@ -40,42 +41,71 @@ namespace WindowsFormsApplication1
         {
             dlgOpen.InitialDirectory = lastPath;
             dlgOpen.DefaultExt = ".map";
-            dlgOpen.ShowDialog();
-            if (dlgOpen.FileNames.Count<string>() > 0)
-                lastPath = ""; Path.GetDirectoryName(dlgOpen.FileNames[0]);
-            foreach (String file in dlgOpen.FileNames)
+            if (dlgOpen.ShowDialog() == DialogResult.OK)
             {
-                cblMain.Items.Add(file);
+                cblMain.Items.Clear();
+                if (dlgOpen.FileNames.Count<string>() > 0)
+                    lastPath = ""; Path.GetDirectoryName(dlgOpen.FileNames[0]);
+                foreach (String file in dlgOpen.FileNames)
+                {
+                    cblMain.Items.Add(file);
+                }
             }
         }
 
-        private void button4_Click(object sender, EventArgs e)
+        private void DoKmz() 
         {
             String filePath;
             String dirPath;
-            for (int i = 0; i < cblMain.CheckedItems.Count; i++)
+            String stIterInfo = "";
+            int i = 0;
+            for (i = 0; i < cblMain.CheckedItems.Count; i++)
             {
-                filePath = (String)cblMain.CheckedItems[i];
-                dirPath = Path.GetDirectoryName(filePath);
-                if (Directory.Exists(dirPath + @"\images"))
-                    Directory.Delete(dirPath + @"\images", true);
-                Directory.CreateDirectory(dirPath+@"\images");
-                File.Copy(dirPath + @"\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg", dirPath + @"\images\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg");
-
-                ProcessStartInfo startInfo = new ProcessStartInfo();
-                startInfo.CreateNoWindow = true;
-                startInfo.UseShellExecute = false;
-                startInfo.FileName = "zip.exe";
-                startInfo.WindowStyle = ProcessWindowStyle.Normal;
-                startInfo.Arguments = "-0 -r \"" + dirPath + "\\" + Path.GetFileNameWithoutExtension(filePath) + ".kmz\" " + "images " + Path.GetFileNameWithoutExtension(filePath) + ".kml";
-                using (Process exeProcess = Process.Start(startInfo))
+                try
                 {
-                    exeProcess.WaitForExit();
+                    filePath = (String)cblMain.CheckedItems[i];
+                    dirPath = Path.GetDirectoryName(filePath);
+                    stIterInfo = i.ToString() + "/" + cblMain.CheckedItems.Count.ToString() + ": " + filePath;
+                    if (!File.Exists(dirPath + @"\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg"))
+                    {
+                        rtbInfo.AppendText("Brak '" + dirPath + @"\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg" + "'" + "\n");
+                        rtbInfo.ScrollToCaret();
+                        continue;
+                    }
+
+                    if (Directory.Exists(dirPath + @"\images"))
+                        Directory.Delete(dirPath + @"\images", true);
+                    Directory.CreateDirectory(dirPath + @"\images");
+                    File.Copy(dirPath + @"\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg", dirPath + @"\images\" + Path.GetFileNameWithoutExtension(filePath) + ".jpg");
+
+                    ProcessStartInfo startInfo = new ProcessStartInfo();
+                    startInfo.CreateNoWindow = true;
+                    startInfo.UseShellExecute = false;
+                    startInfo.FileName = "zip.exe";
+                    startInfo.WindowStyle = ProcessWindowStyle.Normal;
+                    startInfo.Arguments = "-0 -r \"" + dirPath + "\\" + Path.GetFileNameWithoutExtension(filePath) + ".kmz\" " + "images " + Path.GetFileNameWithoutExtension(filePath) + ".kml";
+                    using (Process exeProcess = Process.Start(startInfo))
+                    {
+                        exeProcess.WaitForExit();
+                    }
+                    Directory.Delete(dirPath + @"\images", true);
+                    rtbAkcja.AppendText("Wykonano " + stIterInfo + "\n");
+                    rtbAkcja.ScrollToCaret();
+                    //System.Threading.Thread.Sleep(1000);
+                    this.Refresh();
                 }
-
-
-                Directory.Delete(dirPath + @"\images", true);
+                catch (Exception ex)
+                {
+                    rtbInfo.AppendText("Bład " + stIterInfo + "; " + ex.Message + "\n");
+                    rtbInfo.ScrollToCaret();
+                }
             }
+        }
+
+
+        private void button4_Click(object sender, EventArgs e)
+        {
+            DoKmz();
         }
     }
 }
